requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.0
youtube-transcript-api==0.6.1
langchain==0.1.9
langchain-pinecone==0.0.3
pinecone-client==3.0.2
langchain-groq==0.0.5
langchain-core==0.1.27
langchain-community==0.0.24
sentence-transformers==2.5.1
pydantic==2.6.1